Add tag plugin and switch categories to tags
Tage a tag page generating plugin from the jekyll-clean-dark theme and
use in this project. Fix the corresponding layout file to work with this
page. Also update license and footer accordingly.

Updat the right menu somewhat to prepare for adding tags there.

Fix so that all posts now use the tag system instead of the categories.

Fix a new icon.