Fix cache pages, remove tag plugin
Remove the plugin generating tag pages and just add a template instead.
Markdown on the pages does not work otherwise. When adding a new tag a
folder for that must be copied from template, folder renamed and tag
value in front matter fixed.

Also make sure that the correct place in menu gets 'current'.

Small canges in about page.