Fixed page building bug from jekyll-feed + cleanup
Fixed the an error that caused the page not to build due to
http not being included in the url. This caused the jekyll-feed
plugin to throw an error, and so the page was not built by github
pages properly.

Also added a proper Gemfile, and renamed some posts (cleanup).